refactor(context): move pending-seed machinery into with_event_context
The pending-seed stack, materialization, and poll guard exist solely to
support EventContextFuture's lazy seeding — they are wrapper plumbing,
not part of the EventContext/ContextData core, so they live with the
future code. mod.rs keeps only the two materialize calls at its stack
entry points (current/seed). SeedGuard drops from pub(crate) to
module-private.

No behavior change; poll-path benchmarks unchanged (-40% vs main).
