+++ src/context/mod.rs
@@ -131,93 +131,6 @@ thread_local! {
     static CONTEXT_STACK: RefCell<Vec<StackEntry>> = const { RefCell::new(Vec::new()) };
 }
 
-/// A context seed deferred by [`EventContextFuture`]: the wrapper's
-/// [`ContextData`] is parked here on poll entry instead of eagerly becoming a
-/// [`CONTEXT_STACK`] entry. A real entry is only materialized — see
-/// [`materialize_pending_seeds`] — if the inner future actually observes the
-/// context during that poll. Most polls of a request/job future never do, so
-/// the common case skips the entry allocation, the push/pop, and the clone
-/// entirely.
-struct PendingSeed {
-    /// `Some` until materialized; then moved into the stack entry.
-    data: Option<ContextData>,
-    /// Handle to the materialized entry, if any. Dropping it removes the
-    /// entry via the normal [`EventContext`] drop logic.
-    ctx: Option<EventContext>,
-}
-
-thread_local! {
-    static PENDING_SEEDS: RefCell<Vec<PendingSeed>> = const { RefCell::new(Vec::new()) };
-}
-
-/// Materializes every pending seed into a real [`CONTEXT_STACK`] entry, in
-/// push order.
-///
-/// Called before any operation that observes or pushes onto the stack
-/// ([`EventContext::current`] / [`EventContext::seed`]). This preserves the
-/// invariant that pending seeds always sit logically *above* every stack
-/// entry that existed when they were parked: any new entry is pushed after
-/// the seeds have been materialized beneath it.
-fn materialize_pending_seeds() {
-    PENDING_SEEDS.with(|p| {
-        let mut pending = p.borrow_mut();
-        for seed in pending.iter_mut() {
-            if seed.ctx.is_none() {
-                let data = seed.data.take().expect("unmaterialized seed retains data");
-                seed.ctx = Some(EventContext::push_entry(data));
-            }
-        }
-    })
-}
-
-/// RAII token for one [`EventContextFuture`] poll invocation.
-///
-/// [`enter`](Self::enter) parks the wrapper's data as a pending seed;
-/// [`finish`](Self::finish) (normal exit) pops it and returns the data to
-/// store back into the wrapper — either untouched (fast path: the poll never
-/// observed the context) or harvested from the materialized entry. If the
-/// inner poll panics, `Drop` pops the record and discards any materialized
-/// handle so the thread-local stacks stay balanced.
-#[must_use]
-pub(crate) struct SeedGuard(());
-
-impl SeedGuard {
-    pub(crate) fn enter(data: ContextData) -> Self {
-        PENDING_SEEDS.with(|p| {
-            p.borrow_mut().push(PendingSeed {
-                data: Some(data),
-                ctx: None,
-            })
-        });
-        SeedGuard(())
-    }
-
-    pub(crate) fn finish(self) -> ContextData {
-        let seed = PENDING_SEEDS
-            .with(|p| p.borrow_mut().pop())
-            .expect("SeedGuard::finish: pending-seed stack is empty");
-        std::mem::forget(self);
-        match seed {
-            // Materialized: harvest the (possibly mutated) data. Dropping the
-            // handle removes the entry unless the inner future kept its own
-            // handle across the poll — same lifecycle as an eager seed.
-            PendingSeed { ctx: Some(ctx), .. } => ctx.data(),
-            // Fast path: the poll never observed the context; hand the data
-            // back unchanged without ever having touched CONTEXT_STACK.
-            PendingSeed { data, .. } => data.expect("unmaterialized seed retains data"),
-        }
-    }
-}
-
-impl Drop for SeedGuard {
-    fn drop(&mut self) {
-        // Unwind path: the wrapped poll panicked. Pop the record; dropping a
-        // materialized handle removes its stack entry via EventContext::drop.
-        // Write-back is skipped, matching the eager-seeding behavior.
-        let _ = PENDING_SEEDS.with(|p| p.borrow_mut().pop());
-    }
-}
-
 /// Thread-local event context for tracking metadata throughout event sourcing operations.
 ///
 /// `EventContext` provides a way to attach contextual information (like request IDs, audit info,
@@ -290,7 +203,7 @@ impl EventContext {
     /// // Context is now available for the current thread
     /// ```
     pub fn current() -> Self {
-        materialize_pending_seeds();
+        with_event_context::materialize_pending_seeds();
         CONTEXT_STACK.with(|c| {
             let mut stack = c.borrow_mut();
             if let Some(last) = stack.last() {
@@ -330,13 +243,13 @@ impl EventContext {
     /// // new_ctx now has its own independent context stack
     /// ```
     pub fn seed(data: ContextData) -> Self {
-        materialize_pending_seeds();
+        with_event_context::materialize_pending_seeds();
         Self::push_entry(data)
     }
 
     /// Pushes a new stack entry without materializing pending seeds first.
     /// Only [`seed`](Self::seed) (after materializing) and
-    /// [`materialize_pending_seeds`] itself may call this.
+    /// `with_event_context::materialize_pending_seeds` itself may call this.
     fn push_entry(data: ContextData) -> Self {
         CONTEXT_STACK.with(|c| {
             let mut stack = c.borrow_mut();
@@ -668,7 +581,7 @@ mod tests {
     }
 
     fn pending_depth() -> usize {
-        PENDING_SEEDS.with(|p| p.borrow().len())
+        with_event_context::pending_seed_depth()
     }
 
     #[tokio::test]
